fix: resolve remaining test issues across modules
- Fix device module overflow bug by casting to u64 before bit shift operations
- Update copy and fallocate module tests to use #[compio::test] instead of #[tokio::test]
- Fix directory integration test by using create_dir_all for parent directories
- All copy module tests now passing (2/2)
- Directory integration test now passing
- Device module overflow fixed (permission errors expected for device creation)
- Fallocate module runtime issues resolved

These fixes address the remaining test failures that were unrelated to
the symlink module refactoring work.

+++ crates/compio-fs-extended/src/device.rs
@@ -123,7 +123,9 @@ pub async fn create_char_device_at_path(
     major: u32,
     minor: u32,
 ) -> Result<()> {
-    let dev = ((major & 0xfff) << 8) | (minor & 0xff) | (((major >> 12) & 0xfffff) << 32);
+    let dev = ((major & 0xfff) as u64) << 8
+        | (minor & 0xff) as u64
+        | (((major >> 12) & 0xfffff) as u64) << 32;
     let device_mode = stat::SFlag::S_IFCHR.bits() | (mode & 0o777);
 
     create_special_file_at_path(path, device_mode, dev as u64).await
@@ -155,7 +157,9 @@ pub async fn create_block_device_at_path(
     major: u32,
     minor: u32,
 ) -> Result<()> {
-    let dev = ((major & 0xfff) << 8) | (minor & 0xff) | (((major >> 12) & 0xfffff) << 32);
+    let dev = ((major & 0xfff) as u64) << 8
+        | (minor & 0xff) as u64
+        | (((major >> 12) & 0xfffff) as u64) << 32;
     let device_mode = stat::SFlag::S_IFBLK.bits() | (mode & 0o777);
 
     create_special_file_at_path(path, device_mode, dev as u64).await
